+++ src/OneWare.TypeScript/LanguageServiceTypeScript.cs
@@ -9,7 +9,7 @@ public class LanguageServiceTypeScript : LanguageServiceLspAutoDownload
     public LanguageServiceTypeScript(string workspace, ISettingsService settingsService,
         IPackageService packageService)
         : base(settingsService.GetSettingObservable<string>(TypeScriptModule.LspPathSetting),
-            TypeScriptModule.TsgoPackage, TypeScriptModule.LspName, workspace, packageService,
+            TypeScriptModule.TypeScriptPackage, TypeScriptModule.LspName, workspace, packageService,
             settingsService.GetSettingObservable<bool>("Experimental_AutoDownloadBinaries"),
             arguments: "--lsp -stdio")
     {